fix(ci): improve validate.sh delta check filtering for standard cargo outputs
Updates the `normalize_errors` and `delta_check` functions inside `scripts/validate.sh` to properly exclude typical stdout logging from newer `cargo clippy` and `cargo test` runs (e.g. `Finished dev profile`, `Updating crates.io`, `test ... ok` lines). This prevents these successful completion messages from registering as false-positive failures during the `comm -13` baseline comparison.
